% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_mito_genes.R
\docType{data}
\name{human_mito_genes}
\alias{human_mito_genes}
\title{Human Mitochondrial and Nuclear Localized Gene Names}
\format{
A character vector with gene names.
}
\usage{
human_mito_genes
}
\description{
Character vector of mitochondrial and nuclear localized gene names.
}
\keyword{datasets}
