% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mito_genes.R
\docType{data}
\name{mito_genes}
\alias{mito_genes}
\title{Mitochondrial and Nuclear Gene Names}
\format{
A character vector with 439 gene names.
}
\usage{
mito_genes
}
\description{
A dataset containing mitochondrial and nuclear gene names.
}
\keyword{datasets}
