% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_mito_genes.R
\docType{data}
\name{mouse_mito_genes}
\alias{mouse_mito_genes}
\title{Mouse Gene Names}
\format{
A character vector with gene names.
}
\usage{
mouse_mito_genes
}
\description{
Character vector of mitochondrial and nuclear localized gene names.
}
\keyword{datasets}
