% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_genes.R
\docType{data}
\name{rb_genes}
\alias{rb_genes}
\title{Ribosomal Gene Names}
\format{
A character vector with gene names.
}
\usage{
rb_genes
}
\description{
A dataset containing ribosomal gene names.
}
\keyword{datasets}
